基本bug修改

+++ yyms-framework-model/src/main/java/com/leanin/domain/vo/FormInfoVo.java
@@ -20,7 +20,7 @@ public class FormInfoVo {
 	
     private String formNum;		//表单号
 
-    private Integer formCreater;		//表单创建人
+    private Long formCreater;		//表单创建人
 
     private String createrName;     //创建人名称
